% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzstat_get.R
\name{nzstat_get}
\alias{nzstat_get}
\title{Get data from an Aotearoa Data Explorer dataflow}
\usage{
nzstat_get(
  dataflow_id,
  dimensions = list(),
  keep = "factor",
  keep_status = FALSE,
  max_tries = 10L,
  cache = TRUE,
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{dataflow_id}{String. The DataflowID of a dataflow in the API.}

\item{dimensions}{Named list specifying codes to include for each dimension
of the dataset. Each element should be a character string listing the
codes to include for the corresponding dimension. All codes will be
included for omitted dimensions, so only those dimensions you wish to
subset need be listed. Each name should correspond to one of the elements
in the \code{Name} column of \code{nzstat_get_datastructure(dataflow_id)}.}

\item{keep}{String. Whether to keep dimension \"codes\", \"labels\", or
\"both\"; or \"factor\" to convert the labels to a factor variable. The
default is \"factor\".}

\item{keep_status}{Boolean. Whether to keep the Status column in the returned
dataset. The default is \code{FALSE}, as this is flag is rarely set.}

\item{max_tries}{Integer; maximum retry attempts. Passed to
\code{\link[httr2:req_retry]{httr2::req_retry()}}.}

\item{base_url}{The base URL to the API. If not set, uses the
\code{getOption("NZSTAT_BASE_URL")}, or defaults to
\code{"https://api.data.stats.govt.nz/rest/"}}

\item{api_key}{The API key to authenticate with the server. If not specified,
will look in environment variable \code{"NZSTAT_API_KEY"}, and throw an error
if that is not found.}
}
\value{
A tibble, with columns \code{DimensionID}: the ID used to identify the
data dimension in requests; \code{CodeID}: the ID used to identify code values
for the corresponding dimension; and \code{Name}: the name/description of the
code.
}
\description{
Retrieves data from an Aotearoa Data Explorer dataflow.
}
