% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzstat_get_codelists.R
\name{nzstat_get_codelists}
\alias{nzstat_get_codelists}
\title{Get codelists for dimensions of an Aotearoa Data Explorer dataflow}
\usage{
nzstat_get_codelists(
  dataflow_id,
  dimension_ids,
  max_tries = 10L,
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{dataflow_id}{String. The DataflowID of a dataflow in the API.}

\item{dimension_ids}{Character vector. Dimensions for which to retrieve
codelists.}

\item{max_tries}{Integer; maximum retry attempts. Passed to
\code{\link[httr2:req_retry]{httr2::req_retry()}}.}

\item{base_url}{The base URL to the API. If not set, uses the
\code{getOption("NZSTAT_BASE_URL")}, or defaults to
\code{"https://api.data.stats.govt.nz/rest/"}}

\item{api_key}{The API key to authenticate with the server. If not specified,
will look in environment variable \code{"NZSTAT_API_KEY"}, and throw an error
if that is not found.}
}
\value{
A tibble, with columns \code{DimensionID}: the ID used to identify the
data dimension in requests; \code{CodeID}: the ID used to identify code values
for the corresponding dimension; and \code{Name}: the name/description of the
code.
}
\description{
Retrieves and formats codelists for one or more dimensions of an Aotearoa
Data Explorer dataflow.
}
