% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzstat_get_dataflows.R
\name{nzstat_get_dataflows}
\alias{nzstat_get_dataflows}
\title{Get dataflow definitions for all available Aotearoa Data Explorer datasets}
\usage{
nzstat_get_dataflows(
  search = NULL,
  max_tries = 10L,
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{search}{Optional search string. A regex pattern passed to \code{\link[=grepl]{grepl()}}.}

\item{max_tries}{\code{FALSE}, or integer giving maximum retry attempts. Passed to
\code{\link[httr2:req_retry]{httr2::req_retry()}}.}

\item{base_url}{The base URL to the API. If not set, uses the
\code{getOption("NZSTAT_BASE_URL")}, or defaults to
\code{"https://api.data.stats.govt.nz/rest/"}}

\item{api_key}{The API key to authenticate with the server. If not specified,
will look in environment variable \code{"NZSTAT_API_KEY"}, and throw an error
if that is not found.}
}
\value{
A tibble, with columns \code{Topic}: the broad topic area; \code{Category} the
sub-topic into which the dataflow is categorised; \code{Name}: the
name/description of the dataflow; and \code{DataflowID}: the ID used to
identify the dataflow in requests.
}
\description{
Retrieves, formats, and optionally filters all available dataflow definitions
from the Aotearoa Data Explorer API.
}
