% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzstat_get_datastructure.R
\name{nzstat_get_datastructure}
\alias{nzstat_get_datastructure}
\title{Get the datastructure definition for an Aotearoa Data Explorer dataflow}
\usage{
nzstat_get_datastructure(
  dataflow_id,
  max_tries = 10L,
  base_url = get_base_url(),
  api_key = get_api_key()
)
}
\arguments{
\item{dataflow_id}{String. The DataflowID of a dataflow in the API.}

\item{max_tries}{Integer; maximum retry attempts. Passed to
\code{\link[httr2:req_retry]{httr2::req_retry()}}.}

\item{base_url}{The base URL to the API. If not set, uses the
\code{getOption("NZSTAT_BASE_URL")}, or defaults to
\code{"https://api.data.stats.govt.nz/rest/"}}

\item{api_key}{The API key to authenticate with the server. If not specified,
will look in environment variable \code{"NZSTAT_API_KEY"}, and throw an error
if that is not found.}
}
\value{
A tibble, with column \code{DimensionID}: the ID used to identify the
data dimension in requests.
}
\description{
Retrieves and format the datastructure definition for an Aotearoa Data
Explorer dataflow.
}
