% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_api_key}
\alias{set_api_key}
\title{Store Aotearoa Data Explorer API key in environment variable}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{String: API key. If not specified, \code{\link[askpass:askpass]{askpass::askpass()}} will
be used to provide a prompt to enter the key.}
}
\description{
A helper function to set the \code{NZSTAT_API_KEY} env var.
}
