% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzstat_get_dataflows.R
\name{nzstat_get_dataflows}
\alias{nzstat_get_dataflows}
\title{Get dataflow definitions for all available Aotearoa Data Explorer datasets}
\usage{
nzstat_get_dataflows(search = NULL, max_tries = 10L)
}
\description{
Retrieves, formats, and optionally filters all available dataflow definitions from
}
